Changing project directory & initiating test pub
