--- projects/ros2-tb4-visual-nav/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/ros2-tb4-visual-nav
-[install]
-install_scripts=$base/lib/ros2-tb4-visual-nav
