+++ projects/ros2_tb4_visual_nav/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/ros2_tb4_visual_nav
+[install]
+install_scripts=$base/lib/ros2_tb4_visual_nav
