+++ projects/ros2_tb4_visual_nav/worlds/custom_world.sdf
